Create Change-Child-Support/instructions-faq-about-child-support-in-idaho.md
Created sub-folder Change-Child-Support inside Change-Custody folder. Added instructions-faq-about-child-support-in-idaho.md file inside Change-Child-Support folder. 

These are the instructions if you are changing Child Support in your Modification. They instruct you on how to fill out the other forms included in the Change-Child-Support folder. This is for Modification using the Court Assistance Office of the State of Idaho Judicial Branch.